orii_o1/requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.18.0
accelerate>=0.20.0
tokenizers>=0.13.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0
datasets>=2.12.0
opencv-python>=4.7.0
scikit-learn>=1.2.0

# Web Interface
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=20.1.0
websockets>=11.0.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
rich>=13.0.0
click>=8.1.0
pyyaml>=6.0
jsonlines>=3.1.0
requests>=2.31.0

# Image Processing
torchvision>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.3.0

# Optional GPU acceleration
# triton>=2.0.0  # For faster attention on compatible GPUs